type cast warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@4382 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/TimeKeeper.h
@@ -110,7 +110,7 @@ inline bool		TimeKeeper::operator<=(const TimeKeeper& t) const
 
 inline float		TimeKeeper::getSeconds(void) const
 {
-  return seconds;
+  return (float)seconds;
 }
 
 #endif // BZF_TIME_KEEPER_H
